feat: add proxy impl for TcpProxy, extra docs
