+++ src/proxy.rs
@@ -1,15 +1,14 @@
-use crate::config::{Backend, Config, Protocol, Target};
+use crate::config::{Backend, Config, Protocol};
 use crate::lb::SendTargets;
 use anyhow::Result;
 use async_trait::async_trait;
 use dashmap::DashMap;
 use std::iter::Iterator;
-use std::{collections::HashMap, sync::Arc, vec};
-use tokio::io::AsyncReadExt;
+use std::{sync::Arc, vec};
 use tokio::io::AsyncWriteExt;
 use tokio::net::{TcpListener, TcpStream};
 use tokio::sync::RwLock;
-use tracing::{debug, error, info};
+use tracing::{error, info};
 
 /// Proxy is used to encompass common functionality between L4 and L7 load balancing.
 ///
@@ -17,7 +16,7 @@ use tracing::{debug, error, info};
 /// makes it possible.
 #[async_trait]
 pub trait Proxy {
-    /// Being accepting a particular type of connection.
+    /// Accepting a particular type of connection for configured listeners and targets.
     ///
     /// This is limited to TCP-oriented connections, e.g. TCP and HTTP.
     async fn accept(
@@ -25,9 +24,8 @@ pub trait Proxy {
         current_healthy_targets: Arc<DashMap<String, Vec<Backend>>>,
     ) -> Result<()>;
 
-    // TODO: add another slash here once impl to stop errors.
-    // TODO: think about adding `self` here for connection related info?
-    // After accepting an incoming connection for a target, it should be proxied to a healthy backend.
+    /// Proxy a `TcpStream` from an incoming connection to configured targets, with accompanying
+    /// `Connection` related data.
     async fn proxy(mut connection: Connection, routing_idx: Arc<RwLock<usize>>) -> Result<()>;
 }
 
@@ -43,9 +41,17 @@ pub struct State {
 pub struct Connection {
     pub targets: Arc<DashMap<String, Vec<Backend>>>,
     pub stream: TcpStream,
-    pub client: Option<Arc<reqwest::Client>>,
     pub target_name: String,
+
+    /// An optional HTTP client used to make requests.
+    pub client: Option<Arc<reqwest::Client>>,
+
+    /// An optional HTTP `method` for the connection, for example `GET`.
+    /// This is only used for HTTP connections.
     pub method: Option<String>,
+
+    /// An optional `request_path` for the connection, for example `/api/v1/users`.
+    /// This is only used for HTTP connections.
     pub request_path: Option<String>,
 }
 
@@ -124,9 +130,6 @@ pub async fn health(conf: Arc<Config>, sender: SendTargets) {
                                         target_name: name.clone(),
                                         backend: backend.clone(),
                                     }))
-                                    // This is "removed from the pool" because it is not included in
-                                    // the vector for the next channel transmission, so traffic does not get routed
-                                    // to it.
                                 }
                             }
                         } else {
@@ -143,113 +146,4 @@ pub async fn health(conf: Arc<Config>, sender: SendTargets) {
     } else {
         info!("No targets configured, cannot perform health checks.");
     }
-}
-
-/// Proxy a TCP connection to a range of configured backend servers.
-pub async fn tcp_connection(
-    targets: Arc<DashMap<String, Vec<Backend>>>,
-    target_name: String,
-    routing_idx: Arc<RwLock<usize>>,
-    mut stream: TcpStream,
-) -> Result<()> {
-    if let Some(backends) = targets.get(&target_name) {
-        let backends = backends.to_vec();
-        debug!("Backends configured {:?}", &backends);
-        let backend_count = backends.len();
-
-        if backend_count == 0 {
-            info!("[TCP] No routable backends for {target_name}, nothing to do");
-            return Ok(());
-        }
-
-        // Limit the scope of the index write lock.
-        let backend_addr: String;
-        {
-            let mut idx = routing_idx.write().await;
-            debug!(
-                "[TCP] {backend_count} backends configured for {target_name}, current index {idx}"
-            );
-
-            // Reset index when out of bounds to route back to the first server.
-            if *idx >= backend_count {
-                *idx = 0;
-            }
-
-            backend_addr = format!("{}:{}", backends[*idx].host, backends[*idx].port);
-
-            // Increment a shared index after we've constructed our current connection
-            // address.
-            *idx += 1;
-        }
-
-        info!("[TCP] Attempting to connect to {}", &backend_addr);
-
-        let mut response = TcpStream::connect(&backend_addr).await?;
-        let mut buffer = Vec::new();
-        response.read_to_end(&mut buffer).await?;
-        stream.write_all(&buffer).await?;
-        debug!("TCP stream closed");
-    } else {
-        info!("[TCP] No backend configured");
-    };
-
-    Ok(())
-}
-
-/// Bind to the configured listener ports for incoming TCP connections.
-async fn generate_tcp_listeners(
-    bind_address: String,
-    targets: HashMap<String, Target>,
-) -> Result<Vec<(String, TcpListener)>> {
-    let mut tcp_bindings = vec![];
-
-    for (name, target) in targets {
-        if target.protocol_type() == Protocol::Tcp {
-            let addr = format!("{}:{}", bind_address.clone(), target.listener.unwrap());
-            let listener = TcpListener::bind(&addr).await?;
-            info!("Binding to {} for {}", &addr, &name);
-            tcp_bindings.push((name, listener));
-        }
-    }
-
-    Ok(tcp_bindings)
-}
-
-/// Accept TCP connections by binding to multiple `TcpListener` socket address and
-/// handling incoming connections, passing them to the configured TCP backends.
-pub async fn accept_tcp(
-    bind_address: String,
-    current_healthy_targets: Arc<DashMap<String, Vec<Backend>>>,
-    targets: HashMap<String, Target>,
-) -> Result<()> {
-    let idx: Arc<RwLock<usize>> = Arc::new(RwLock::new(0));
-    let bound_listeners = generate_tcp_listeners(bind_address, targets).await?;
-
-    for (name, listener) in bound_listeners {
-        // Listen to incoming traffic on separate threads
-        let idx = Arc::clone(&idx);
-        let current_healthy_targets = Arc::clone(&current_healthy_targets);
-
-        tokio::spawn(async move {
-            while let Ok((stream, remote_peer)) = listener.accept().await {
-                info!("Incoming request on {remote_peer}");
-
-                let idx = Arc::clone(&idx);
-                let tcp_targets = Arc::clone(&current_healthy_targets);
-                // Pass the TCP streams over to separate threads to avoid
-                // blocking and give each thread its copy of the configuration.
-                let target_name = name.clone();
-
-                tokio::spawn(async move {
-                    tcp_connection(tcp_targets, target_name, idx, stream)
-                        .await
-                        .unwrap();
-                })
-                .await
-                .unwrap();
-            }
-        });
-    }
-
-    Ok(())
 }
